Updated lift sources to work with new eval API returning a Long or call alternative method which returns an Int.

+++ src/main/c/generator/CAst.scala
@@ -50,7 +50,7 @@ object CAst {
                      t: Type,
                      init: OclAstNode = null,
                      addressSpace: OpenCLAddressSpace = UndefAddressSpace,
-                     length: Int = 0) extends Declaration
+                     length: Long = 0) extends Declaration
 
   /** Parameter declaration. These have to be separated from variable
     * declaration since the vectorization has to be handled differently
